Refactor classes in file to separate package
