+++ src/nl/youngcapital/kermis/attracties/Botsautos.java
@@ -1,0 +1,8 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class Botsautos extends Attractie{
+   public Botsautos(){
+      attractieNaam = "botsautos";
+   }
+}
