+++ src/nl/youngcapital/kermis/attracties/Hawaii.java
@@ -1,0 +1,9 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class Hawaii extends Attractie{
+   String attractieNaam = "hawaii";
+   public void attractieDraaien() {
+      System.out.println("De attractie " +  attractieNaam + " draait.");
+   }
+}
