+++ src/nl/youngcapital/kermis/attracties/LadderKlimmen.java
@@ -1,0 +1,9 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class LadderKlimmen extends Attractie{
+   String attractieNaam = "ladderklimmen";
+   public void attractieDraaien() {
+      System.out.println("De attractie " +  attractieNaam + " draait.");
+   }
+}
