+++ src/nl/youngcapital/kermis/attracties/Spiegelpaleis.java
@@ -1,0 +1,9 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class Spiegelpaleis extends Attractie{
+   String attractieNaam = "spiegelpaleis";
+   public void attractieDraaien() {
+      System.out.println("De attractie " +  attractieNaam + " draait");
+   }
+}
