+++ src/nl/youngcapital/kermis/attracties/Spin.java
@@ -1,0 +1,9 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class Spin extends Attractie{
+   String attractieNaam = "spin";
+   public void attractieDraaien() {
+      System.out.println("De attractie " +  attractieNaam + " draait!");
+   }
+}
