+++ src/nl/youngcapital/kermis/attracties/Spookhuis.java
@@ -1,0 +1,9 @@
+package nl.youngcapital.kermis.attracties;
+
+
+public class Spookhuis extends Attractie{
+   String attractieNaam = "spookhuis";
+   public void attractieDraaien() {
+      System.out.println("De attractie " +  attractieNaam + " draait.");
+   }
+}
